Adicionando SlopeOne

+++ app/src/main/java/mpoo/bsi/ufrpe/helpvegapp/avaliacao/domain/Rating.java
@@ -7,10 +7,10 @@ public class Rating {
     private int ratingId;
     private User userRating;
     private Restaurant restaurantRating;
-    private float food;
-    private float price;
-    private float service;
-    private float ambiance;
+    private double food;
+    private double price;
+    private double service;
+    private double ambiance;
 
     public int getRatingId() {
         return ratingId;
@@ -33,35 +33,35 @@ public class Rating {
         this.restaurantRating = restaurantRating;
     }
 
-    public float getFood() {
+    public double getFood() {
         return food;
     }
-    public void setFood(float food) {
+    public void setFood(double food) {
         this.food = food;
     }
 
-    public float getPrice() {
+    public double getPrice() {
         return price;
     }
-    public void setPrice(float price) {
+    public void setPrice(double price) {
         this.price = price;
     }
 
-    public float getService() {
+    public double getService() {
         return service;
     }
-    public void setService(float service) {
+    public void setService(double service) {
         this.service = service;
     }
 
-    public float getAmbiance() {
+    public double getAmbiance() {
         return ambiance;
     }
-    public void setAmbiance(float ambiance) {
+    public void setAmbiance(double ambiance) {
         this.ambiance = ambiance;
     }
 
-    public float getGeneral(){
+    public double getGeneral(){
         return (getAmbiance()+getFood()+getPrice()+getService()) / 4;
     }
 }
